chore: build fix
