+++ packages/toolkit/src/view/dashboard/DashboardPipelineListPageMainView.tsx
@@ -14,6 +14,10 @@ import {
   useTriggeredPipelines,
   useTriggeredPipelinesChart,
   dashboardOptions,
+  useUser,
+  useNamespaceType,
+  useUsersSubscription,
+  useOrganizationsSubscription,
 } from "../../lib";
 import { FilterByDay } from "./FilterByDay";
 import { PipelineTriggerCountsLineChart } from "./PipelineTriggerCountsLineChart";
@@ -139,6 +143,56 @@ export const DashboardPipelineListPageMainView = (
     triggeredPipelineList,
   ]);
 
+  // Remaing Triggers
+
+  const instillUser = useUser({
+    enabled: enableQuery || false,
+    retry: false,
+    accessToken,
+  });
+
+  const namespaceType = useNamespaceType({
+    namespace: String(instillUser.data?.id) ?? null,
+    enabled: enableQuery,
+    accessToken,
+  });
+
+  const userSubscription = useUsersSubscription({
+    userName: instillUser.isSuccess ? instillUser.data.id : null,
+    enabled:
+      namespaceType.isSuccess && namespaceType.data === "NAMESPACE_USER"
+        ? enableQuery
+        : false,
+    accessToken,
+  });
+  const organizationSubscription = useOrganizationsSubscription({
+    oraganizationName: instillUser.isSuccess ? instillUser.data.id : null,
+    enabled:
+      namespaceType.isSuccess && namespaceType.data === "NAMESPACE_ORGANIZATION"
+        ? true
+        : false,
+    accessToken,
+  });
+
+  const subscriptions = React.useMemo(() => {
+    if (namespaceType.data === "NAMESPACE_USER") {
+      if (userSubscription.isLoading) {
+        userSubscription.refetch();
+      }
+      if (userSubscription.data) {
+        return userSubscription;
+      }
+    }
+    if (namespaceType.data === "NAMESPACE_ORGANIZATION") {
+      if (organizationSubscription.isLoading) {
+        organizationSubscription.refetch();
+      }
+      if (organizationSubscription.data) {
+        return organizationSubscription;
+      }
+    }
+  }, [namespaceType, userSubscription, organizationSubscription]);
+
   /* -------------------------------------------------------------------------
    * Render
    * -----------------------------------------------------------------------*/
@@ -166,7 +220,11 @@ export const DashboardPipelineListPageMainView = (
                 pipelineTriggersSummary ? pipelineTriggersSummary.errored : null
               }
             />
-            <RemainingTriggers />
+            <RemainingTriggers
+              subscriptions={
+                subscriptions?.isSuccess ? subscriptions?.data : null
+              }
+            />
           </PipelineTriggersSummary>
         </div>
